pyproject.toml
[project]
name = "dub"
version = "0.1.0"
description = "AI video dubbing pipeline"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.134.0",
    "fish-audio-sdk>=1.2.0",
    "google-genai>=1.65.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "redis>=7.2.1",
    "sse-starlette>=3.3.2",
    "taskiq>=0.12.1",
    "taskiq-redis>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "poethepoet>=0.42.1",
]

[tool.poe.tasks]
api = "uvicorn dub.main:app --reload"
worker = "taskiq worker dub.worker:broker"
